+++ src/application/deployer/push_image_options.go
@@ -1,0 +1,18 @@
+package deployer
+
+import (
+	"github.com/Originate/exosphere/src/types"
+	"github.com/aws/aws-sdk-go/service/ecr"
+)
+
+// PushImageOptions is the options to PushServiceImage
+type PushImageOptions struct {
+	DeployConfig     types.DeployConfig
+	DockerComposeDir string
+	EcrAuth          string
+	EcrClient        *ecr.ECR
+	ImageName        string
+	ServiceLocation  string
+	ServiceRole      string
+	BuildImage       bool
+}