+++ src/cmd/deploy.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/Originate/exosphere/src/application"
+	"github.com/Originate/exosphere/src/application/deployer"
 	"github.com/Originate/exosphere/src/config"
 	"github.com/Originate/exosphere/src/docker/composebuilder"
 	"github.com/Originate/exosphere/src/types"
@@ -62,10 +62,10 @@ var deployCmd = &cobra.Command{
 			DeployServicesOnly: deployServicesFlag,
 
 			// git commit hash of the Terraform modules in Originate/exosphere we are using
-			TerraformModulesRef: "d5c193c5",
+			TerraformModulesRef: "01891552",
 		}
 
-		err = application.StartDeploy(deployConfig)
+		err = deployer.StartDeploy(deployConfig)
 		if err != nil {
 			log.Fatalf("Deploy failed: %s", err)
 		}