+++ src/config/rds_production_dependency.go
@@ -51,3 +51,8 @@ func (r *rdsProductionDependency) GetDeploymentServiceEnvVariables(secrets types
 		r.config.Config.Rds.ServiceEnvVarNames.Password: secrets[r.config.Config.Rds.PasswordSecretName],
 	}
 }
+
+// GetDeploymentVariables returns a map from string to string of variables that a dependency Terraform module needs
+func (r *rdsProductionDependency) GetDeploymentVariables() (map[string]string, error) {
+	return map[string]string{}, nil
+}