+++ src/config/service_config.go
@@ -87,7 +87,7 @@ func GetServiceContexts(appContext types.AppContext) (map[string]types.ServiceCo
 	result := map[string]types.ServiceContext{}
 	for service, serviceData := range appContext.Config.GetServiceData() {
 		if len(serviceData.Location) > 0 {
-			serviceContext, err := types.GetServiceContext(appContext, serviceData.Location)
+			serviceContext, err := appContext.GetServiceContext(serviceData.Location)
 			if err != nil {
 				return result, err
 			}